requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.7.1
requests==2.31.0
python-dateutil==2.9.0.post0
